refactor: implementation using the singleton pattern
