--- app/src/main/java/models/ioformatter/base/FormatInput.java
@@ -1,10 +1,0 @@
-package models.ioformatter.base;
-
-import org.checkerframework.checker.nullness.qual.NonNull;
-
-import java.util.List;
-
-public interface FormatInput {
-    @NonNull
-    List<String> formatInput (@NonNull String input);
-}
