--- app/src/main/java/models/ioformatter/base/Formatter.java
@@ -1,3 +1,0 @@
-package models.ioformatter.base;
-
-public abstract class Formatter implements FormatInput, FormatOutput {}
