+++ app/src/main/java/utilities/FormatIngredients.java
@@ -1,23 +1,24 @@
-package models.ioformatter;
+package utilities;
 
-import models.ioformatter.base.Formatter;
 import org.checkerframework.checker.nullness.qual.NonNull;
+import utilities.factory.IngredientsFormatter;
+
 import java.util.Arrays;
 import java.util.List;
 
-public class FormatIngredients extends Formatter {
+public class FormatIngredients implements IngredientsFormatter {
 
-    @Override
     @NonNull
-    public List<String> formatInput (@NonNull String ingredients){
+    @Override
+    public List<String> formatInput(@NonNull String ingredients) {
         return Arrays.asList(ingredients
                 .replaceAll("(\\r?\\n)+", "\n")
                 .split("\n"));
     }
 
-    @Override
     @NonNull
-    public String formatOutput (@NonNull List<String> ingredients){
+    @Override
+    public String formatOutput(@NonNull List<String> ingredients) {
         return String.join("\n", ingredients);
     }
 }
